Crate sui_single_node_benchmark

Modules

command

workload

Functions

run_benchmark

Benchmark a given workload on a specified component. The different kinds of workloads and components can be found in command.rs. \checkpoint_size represents both the size of a consensus commit, and size of a checkpoint if we are benchmarking the checkpoint.