sui_single_node_benchmark

Function run_benchmark

pub async fn run_benchmark(
    workload: Workload,
    component: Component,
    checkpoint_size: usize,
    print_sample_tx: bool,
)

Benchmark a given workload on a specified component. The different kinds of workloads and components can be found in command.rs. \checkpoint_size represents both the size of a consensus commit, and size of a checkpoint if we are benchmarking the checkpoint.